+++ CHANGELOG.md
@@ -1,4 +1,13 @@
+## [1.0.2] - 2023-05-08
+
+### Added
+- `cssBgSupportStyle (Boolean) [false]` - by default it's set to false, but if you set `--cssBgSupport true` then **cssBgSupportStyle** by default is also **true** and if you need you can disable it with `--cssBgSupportStyle false`
+
+### Changed
+- addViews changed to cssBgSupport
+- fixed css bg support logic
+
 ## [1.0.1] - 2023-05-07
 
 ### Added
-- addViews parameter (default: false) - adds view and use to svg for use in css
+- addViews parameter (Boolean) [false] - adds view and use to svg for use in css